Move some helpers where they belong
