+++ frontend/exporter/src/traits.rs
@@ -94,9 +94,78 @@ pub mod rustc {
         }
     }
 
+    #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
+    pub struct AnnotatedPredicate<'tcx> {
+        pub is_extra_self_predicate: bool,
+        /// Note: they are all actually `Clause`s.
+        pub predicate: Predicate<'tcx>,
+        pub span: rustc_span::Span,
+    }
+
+    #[extension_traits::extension(pub trait TyCtxtExtPredOrAbove)]
+    impl<'tcx> TyCtxt<'tcx> {
+        /// Just like `TyCtxt::predicates_defined_on`, but in the case of
+        /// a trait or impl item, also includes the predicates defined on
+        /// the parent.
+        fn predicates_defined_on_or_above(
+            self,
+            did: rustc_span::def_id::DefId,
+        ) -> Vec<AnnotatedPredicate<'tcx>> {
+            let mut next_did = Some(did);
+            let mut predicates = vec![];
+            while let Some(did) = next_did {
+                let (preds, parent) = self.annotated_predicates_of(did);
+                next_did = parent;
+                predicates.extend(preds)
+            }
+            predicates
+        }
+
+        fn annotated_predicates_of(
+            self,
+            did: rustc_span::def_id::DefId,
+        ) -> (
+            impl Iterator<Item = AnnotatedPredicate<'tcx>>,
+            Option<rustc_span::def_id::DefId>,
+        ) {
+            let with_self = self.predicates_of(did);
+            let parent = with_self.parent;
+            let with_self = {
+                let extra_predicates: Vec<(Clause<'_>, rustc_span::Span)> =
+                    if rustc_hir::def::DefKind::OpaqueTy == self.def_kind(did) {
+                        // An opaque type (e.g. `impl Trait`) provides
+                        // predicates by itself: we need to account for them.
+                        self.explicit_item_bounds(did)
+                            .skip_binder() // Skips an `EarlyBinder`, likely for GATs
+                            .iter()
+                            .copied()
+                            .collect()
+                    } else {
+                        vec![]
+                    };
+                with_self.predicates.iter().copied().chain(extra_predicates)
+            };
+            let without_self: Vec<Clause<'_>> = self
+                .predicates_defined_on(did)
+                .predicates
+                .iter()
+                .copied()
+                .map(|(clause, _)| clause)
+                .collect();
+            (
+                with_self.map(move |(clause, span)| AnnotatedPredicate {
+                    is_extra_self_predicate: !without_self.contains(&clause),
+                    predicate: clause.as_predicate(),
+                    span,
+                }),
+                parent,
+            )
+        }
+    }
+
     // FIXME: this has visibility `pub(crate)` only because of https://github.com/rust-lang/rust/issues/83049
     pub(crate) mod search_clause {
-        use super::{Path, PathChunk};
+        use super::{AnnotatedPredicate, Path, PathChunk, TyCtxtExtPredOrAbove};
         use crate::rustc_utils::*;
         use rustc_middle::ty::*;
 
@@ -397,7 +466,6 @@ pub mod rustc {
             }
             .with_args(impl_exprs(tcx, owner_id, &nested)?, *tref),
             ImplSource::Param(nested) => {
-                use crate::TyCtxtExtPredOrAbove;
                 let predicates = tcx.predicates_defined_on_or_above(owner_id);
                 let Some((path, apred)) =
                     search_clause::path_to(tcx, &predicates, tref.clone(), param_env)
